// src/lib/schemas/bgm-index-api.d.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
    "/api/index/{index_id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Get a index by id */
        get: operations["get_IndexFetch"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: never;
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    get_IndexFetch: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                /** @description Index id */
                index_id: number;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description Returns a single index if found */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        index: {
                            bgm_id: number;
                            name: string;
                            description?: string;
                            subject_ids: number[];
                            character_ids: number[];
                            person_ids: number[];
                        };
                    };
                };
            };
            /** @description Index not found */
            404: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        title: string;
                        details: {
                            path: string;
                            method: string;
                        };
                        description: string;
                    };
                };
            };
        };
    };
}